[panel] Use randr 1.3 features to detect the first monitor when possible
With randr 1.3, it's possible to set a primary monitor. So we now check
if one is set.

Also, the detection of laptop screens is now improved thanks to the new
connector type property: it's set to "Panel" in this case.

With all those various methods to detect when an output should be first,
we now have a function dedicated to this.