[clock] Update the time when the user clicks to open calendar
When the time in the clock is not right, the user might want to check
what's going on by clicking on it. It's not a big cost to update the
time if necessary there.

A new function refresh_click_timeout_time_only() to only update the time
if necessary, instead of updating the time format, requesting a new
size, etc. is created and used in other parts of the code where
refresh_clock_timeout() was used before.

Closes: https://bugzilla.novell.com/show_bug.cgi?id=507316