[panel] Always show icons for categories in the applications menu
After feedback from distributors, users and people from the usability
team, it appears that it's better to keep the icons there. The rationale
is that the categories are dynamic since they can appear/disappear
depending on the applications you install.

Note that Preferences/Administration in the System menu don't have icons
because they are static.