[clock] Use icon fallbacks if the theme does not include phases of moon
The changes for bgo#564958 (gweather doesn't show phase of moon) will
allow weather_info_get_icon_name() to return values that aren't expected
to be provided in all themes.

We now use the GTK_ICON_LOOKUP_GENERIC_FALLBACK flag so that the
application will fallback to a full moon if a shaded moon isn't
available.

https://bugzilla.gnome.org/show_bug.cgi?id=603306