[panel] Port away from GnomeIconEntry
We add a new PanelIconChooser widget, which is a simple button opening a
GtkFileChooser.

Some code is also factored in libpanel-util:
 - API to easily add an image preview to a file chooser (moved away from
   gnome-panel-properties.c)
 - some XDG-related icon theme API (moved away from panel-util.c:
   panel_util_icon_remove_extension is renamed to
   panel_xdg_icon_remove_extension)

The new PanelIconChooser is now used in panel-ditem-editor.c and
panel-properties-dialog.c. This enables us to remove some workaround to
handle icon themes in a hacky way -- this is now hidden in
PanelIconChooser.