[panel] PanelAppletFrame needs to be a subclass of GtkEventBox
It was a subclass of GtkBin, which made it impossible to catch
right-click button events on the handles.

This also fixes the annoying bug where applets are not always visible
(this was an issue caused by the way the allocation was used when an
applet has a handle -- apparently, GtkBin doesn't work as I thought in
this case).