[panel] Fix a crash when an applet cannot be loaded
We load applets asynchronously, so we specifically don't call
panel_applet_stop_loading() for this type. However, in case of failure
during the load, we might call panel_applet_stop_loading()
synchronously, which will make us lose the content of the applet
variable. So we save the type to be sure we always ignore the applets.