[panel] Do not forget to call panel_applet_stop_loading() on issues
When we don't load an applet because it's on the no-reload list, because
it's disabled in by lockdown keys or because we can't find the
compatibility iid, we still have to call panel_applet_stop_loading().

If we don't do so, new panels will not get unhidden, since they will
wait for this applet to stop loading -- while it didn't even start
loading.