[panel] Do not limit the initial unhide timeout to initial load
There's no reason to limit the initial unhide timeout of toplevels to
the initial load: a panel created later (with a script that would also
populate it with applets) could also have applets that fail to load. So
we need a timeout there too.

Instead, just don't add the timeout if we already have one.