Prevent modules from being unloadable
The bonobo compatibility module doesn't support unloading.
In general, it's really hard to write a module that can be
safely unloaded and then reloaded later. Also, there's not
much point in supporting it for the panel's use case.  Therefore,
we won't anymore.  That solves the bonobo issue and any future
issues that could potentially crop up.

http://bugzilla.gnome.org/show_bug.cgi?id=629535