Removed gdk_xid_table_lookup_for_display()
According to the documentatiom, GTK+ only stores windows
in its X id table nowadays and the recomendation is to use
gdk_x11_window_lookup_for_display() which was already in
use.  Hence, if the returned value is != NULL then is a window
and we can ask to the user to kill the application.



Signed-off-by: Germán Póo-Caamaño <gpoo@gnome.org>