bonobo: Drop this, it will be moved to a new module if maintained
With the gtk3 branch merged in, it makes no sense to keep the bonobo
support in the gnome-panel module itself.

It's worth keeping the bonobo support in some way, I believe, but it
should be in an external module. Note that some work is needed: we need
a small server that would sit between gnome-panel and the bonobo applets
to make sure that the gnome-panel process doesn't use gtk+ 2.x at all.