panel: Fix dbus applets to work in multiscreen environment
When the applet is created, the frame itself is not yet added to a
widget hierarchy, and is therefore not associated with any screen. So we
always sent the default screen as the one to be used to the applet.

We can simply use the screen of the PanelWidget that will contain the
applet instead.

https://bugzilla.gnome.org/show_bug.cgi?id=632369