panel: Support adding launchers (and menu buttons) from gnomecc.menu
Even though we don't want to expose the control center menu by default,
some advanced users might want to add launchers for some specific
control center panels, so support this in the Addto dialog. Note that,
like in the gnome-control-center shell, we do not display entries that
are directly at the root level of gnomecc.menu.

This also means we can have gnomecc: menu buttons, so support this.