panel: Make edit actions of panel objects a second different popup menu
In the same way we only allow moving the panel when a mouse button
modifier is used (like in metacity), we only display "edit" actions of
panel objects when the modifier is pressed too.

This means a menu with "Remove"/"Move"/"Lock" options is displayed only
when Alt (default setting) is pressed. When this is not the case, the
action menu from the object is displayed instead.