Bug 633566 - fix missing calendar sources, harder.
Still need to do this by querying for type/capabilities rather than
having incestuous knowledge in the client about what the various backend
names are for. But this is at least moving it closer to the right place,
and making it display events from unknown back ends instead of completely
ignoring them.