panel: Talk to gnome-screensaver via dbus
We were using gnome-screensaver-command before, but there's no reason to
stay this way.

This means we lose support for xscreensaver; it could be added back,
if some people want it, but this doesn't make a lot of sense since
xscreensaver doesn't completely integrate in GNOME (in the system
settings, for example).

Also, open the system settings on the screen panel for screensaver
properties.