panel: Fix moving of panel objects to use switch mode again by default
We were looking at the modifier on the panel, and this was set to Alt,
since this is required to open the context menu. So we always used the
free move mode by default, and this could easily break in weird ways
later on. And we really want the switch mode.

So instead, take the modifier from the event triggering the move.

https://bugzilla.gnome.org/show_bug.cgi?id=646041