panel: Fix size of applets using size hints if they are not packed alone
This is for the expanded mode. For some reason I ignore, we were
starting by giving the minimum size to such applets, while we do have
many passes that allow us to move from the maximum size to a smaller
size -- we were instead going from the minimum size to a bigger size.

This change makes sure we really give as much as possible to such
applets. This is needed with the new packing model where the packing
actually just gave the minimal size to size hints applets because they
weren't requesting more.