panel: Update the size hints array used by toplevel at the right times
We were updating this array on size_request() of PanelWidget, erasing
the old array. However, due to how the size_request() now work, it could
happen that this were doing this another time, after the size_request()
of PanelToplevel which was filling the array with the appropriate sizes.

So instead, only update the array when needed (after moving an object,
or when an object gets its size hints set).