panel, libpanel-applet: Correctly test for modifier keys
Using Alt to edit the panels doesn't always work; this is likely because
we were not using gtk_accelerator_get_default_mod_mask() to test the
modifier key.

Note that we double-check that the metacity modifier is in the mask
returned by gtk_accelerator_get_default_mod_mask(), to make sure it
always work -- if it's not there, we default to MOD1.

https://bugzilla.gnome.org/show_bug.cgi?id=649971