panel: Change default size of icon in menu bar & user menu applets to 16
This makes it possible to make the panel smaller than it used to be (the
size of the icon was the blocker) and it also makes the icon consistent
with the notification area icons.

Note that this size doesn't apply to the user menu when it appears in
the main menu (as opposed to the user menu applet).

This can be changed in settings.ini (see GtkSettings doc) with:
 gtk-icon-sizes = panel-menu-bar=16,16

This used to be called panel-foobar, but that name was horrible.

https://bugzilla.gnome.org/show_bug.cgi?id=660846
https://bugzilla.gnome.org/show_bug.cgi?id=661790