panel: Fix themed background pattern
When the panel background is set to None (ie, use system theme), and the
theme defines a background-image css style for the PanelWidget (with
-gtk-gradient, for example), a pattern is created (the default_pattern
member) which is incorrectly drawn in certain areas. The incorrectly
drawn areas of the panel are regions that are not occupied by child
widgets. These regions are wiped by GDK (in
gdk_window_clear_backing_region) to the background pattern previously
set (in panel_background_prepare).

Although it appears as though no pattern is drawn, in fact the gradient
pattern *is* drawn, but not scaled properly so that appears as if only
one gradient value is used to fill.

This patch correctly scales the default_pattern (if present) to the
width & height of the panel window.

https://bugzilla.gnome.org/show_bug.cgi?id=663397