panel: Fix position re-shuffling on orientation change
The ButtonWidget widgets of launchers were accidentally triggering
'EEEEK' code in panel_widget_update_positions.

This can happen when the parent (the PanelWidget) allocation is stale or
uninitialized. For example, when a panel's orientation is changed (eg,
via the Properties dialog) and the toplevel initiates a resize, the
children are queried for preferred height/width. At this point, the
allocations are stale and reflect the values from the previous
orientation.

Instead, use the toplevel's specified size as the size of the
ButtonWidget (accounting for padding in the major direction).

https://bugzilla.gnome.org/show_bug.cgi?id=671328