panel: Fix random grey box on logout
On session logout, a grey box briefly appears in the upper left corner
of the screen. This grey box appears because the PanelAppletFrame is
attempting to notify the user that an applet has died and ask if the
user would like to restart.

This behavior is triggered by an "applet-broken" signal. This signal is
generated when, as widgets are being unrealized & disposed from the
toplevel in panel_shell_quit(), the socket created by
PanelAppletContainer to track PanelApplet plugs is unrealized which
causes the socket to generate a "plug-removed" signal.

Don't generate "applet-broken" signal if the socket is unrealized.

https://bugzilla.gnome.org/show_bug.cgi?id=672111

Signed-off-by: Peter Hurley <peter@hurleysoftware.com>