na: apply style after realize
Under Fedora 18 (GTK-3.6.2), notification icons such as the battery/power
icon from gnome-settings-daemon are appearing all-black on top of a black
background.

This is because style_updated is being called before the tray applet
is realized, meaning that the following code in
na_tray_applet_style_updated() causes a bail-out:

  if (!applet->priv->tray)
    return;

and the style is never applied.

Fix this by updating the style explicitly after realizing the widget.