applet: fix menu location on multi-monitor setup
The main menu should open upward if the applet is in the bottom half of the monitor,
and downward if the applet is in the upper half of the monitor (similarly for left/right).
Calculating the position based on screen size results in wrong placement on setups where
monitors have different size or rotation. Therefore the menu should open towards the middle
of the monitor not the screen.

https://bugzilla.gnome.org/show_bug.cgi?id=614572