simplify background handling
All this code is not needed and has not worked like it should.

GTK+ 3.16 will deprecate some functions used by this code. Setting
background-image with gtk_style_properties_set was not implemented.
See bug: https://bugzilla.gnome.org/show_bug.cgi?id=710184

This restores ability to have transparent panels. Tested with both
in process and out of process applets, with compositing manager and
without it.

If some applets are not transparent that most likely is theme
problem.