free-the-fish.c: Add a gdbus connection and listen to changes of the lockscreen
We draw the fish with cairo directly on the screen. When the lock screen
is activated by the screensaver while the fish is active they end up
drawing over each other. The result is that we see a flickering fish
in the area where the fish is drawing over the lock screen.

This patch adds a dbus connection using the GDBus skeletton generated
from the org.gnome.ScreenSaver dbus interface and connects to the 'active-changed'
signal of the screensaver. When the screen saver is activated the fish
hides itself such that it does not draw over the lock screen.