xstuff: fix cropped pixbuf rendering during zoom animation
This patch fixes the zoom range for the launcher zoom animation.

The destination popup window size is computed based on the width
and height of widget allocation * ZOOM_FACTOR, while the zoom range is
based on width only. When width is higher than height, the zoomed icon
does not fit in the destination popup window and is cropped.

https://bugzilla.gnome.org/show_bug.cgi?id=771166

Signed-off-by: Yannick Gicquel <ygicquel@gmail.com>