theme: make the click event on the task list button work on the edges too
By making the button take all the available space and only styling the
box inside a button, it is possible to have a rounded look with a border
and a margin but still catch the click events on the edges of the panel.

Thanks Alberts Muktupāvels for the idea.

https://bugzilla.gnome.org/show_bug.cgi?id=769012