data: add new settings for icon sizes
Add menu-icon-size and panel-max-icon-size settings that will be
used by applets to allow user control icon size on panels.

menu-icon-size setting will be used for icons in menus and is
intended to replace existing panel-menu setting.

panel-max-icon-size setting will be used to limit icon size and
is intended to replace existing panel-menu-bar setting.