panel-run-dialog: fix endless loop in remove_parameters
https://standards.freedesktop.org/desktop-entry-spec/desktop-entry-spec-latest.html#exec-variables

According to specification double %% are used to escape literal
percentage character. In this case first percentage is removed and
then strstr returns pointer to literal percentage. If this percentage
+ next alpha character does not make recognized field code we enter
endless loop.

Rewrite function to properly handle literal percentage character and
at the same time rename function to remove_field_codes as it seems
more appropriate name for what it is doing.

https://gitlab.gnome.org/GNOME/gnome-panel/issues/17