clock: adapt for data changes in libgweather 3.31.91
Starting with libgweather commit d7682676ac92f1b8, not all stations have
timezone information associated with them. For example, this is the case
for all stations in Russia. In this case, gnome-panel crashed because
wtz was NULL.

To fix that, we find the city nearest to the station, which should have
the timezone information.