libgnome-panel: add popup menu helper function
When applets want to popup a menu for a widget they need to know the
current position of the applet. Then they must compute the widget and
menu anchors that are needed by gtk_menu_popup_at_widget() to align
the widget and the menu.

To avoid that every applet needs to perform these steps we introduce
a helper function that automatically computes the archor parameters
from the applet position and pops up the menu for the widget.