properties-dialog: show list of applets
This adds a new tab to the gnome-panel properties
dialog which is titled 'Applets' and which shows
the list of currently added applets for this panel.

The list of applets is grouped into the three
locations on the panel (left, right, center) and
its possible to remove applets from the panel.

Additionally a kebab-menu (three dots) allows
access to 'About' and 'Help' dialogs for each
applet if they are implemented by that applet.