menu: only consume mouse button 3 press events
According to Alberts, the purpose of hooking the mouse button press
event is "...to not show applet menu if you right click on some menu item."

Nevertheless for hierarchical menus it's important that button 1 presses
propagate down the window hierarchy. Otherwise it's impossible to open
the submenus without a mouse -- the tap on a touch screen seems to
cause the button press event to be emitted.

Let's propagate the events from buttons other than button 3.

https://gitlab.gnome.org/GNOME/gnome-panel/-/merge_requests/51